refac: 独自式の定義を移動
